juggle_challenge/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-extensions==3.1.3
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
flake8==3.9.2
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.8.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
regex==2021.7.6
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
tomli==1.0.4
uritemplate==3.0.1
urllib3==1.26.6
simplejson==3.17.2
